rados: use size_t for `max_obj_len`
otherwise, on some platforms where size_t is not uint64_t, MacOS for
instance, no matched generate_random_number() can be selected.

+++ src/tools/rados/rados.cc
@@ -549,7 +549,7 @@ public:
   int read_percent;
   int num_objs;
   size_t min_obj_len;
-  uint64_t max_obj_len;
+  size_t max_obj_len;
   size_t min_op_len;
   size_t max_op_len;
   size_t max_ops;
